Add waiting for entities and easy event setup
Currently the only (public) usage of this is the next_message() method,
which allows you to wait until a specified timeout for a message that
matches a predicate.

Also, instead of having to manually set up events, now you can just call
setup_event_handlers in the service with your cluster, which will do it
for you. Only saves you like 1 line of code but hey it's something :D